Fix: Extract AbstractConfigTestCase
